backend/app/requirements.txt
Django==6.0.2
psycopg[binary]==3.2.10
djangorestframework==3.16.1
drf-spectacular==0.29.0
Pillow==10.4.0
